project: move rust-analyzer specific code into a dedicated module.
